# docs/openapi/stigmem.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Reference node implementing the Stigmem v0.9.0a1 HTTP API \u2014 facts, federation, gardens, recall, subscriptions, audit, identity, content-addressed fact IDs. Cross-cutting features (tombstones, time-travel, multi-tenant) are opt-in plugins.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Stigmem Reference Node",
    "version": "0.9.0a1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/decay/jobs/{job_id}": {
      "get": {
        "description": "Poll the status of an async decay job (Spec-X9-Decay-Semantics).",
        "operationId": "get_decay_job_v1_decay_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Decay Job V1 Decay Jobs  Job Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Decay Job",
        "tags": [
          "decay"
        ]
      }
    },
    "/v1/decay/sweep": {
      "post": {
        "description": "Mark stale facts as expired. Cron-friendly one-shot sweeper.\n\nReturns 200 synchronously for scopes \u2264 threshold facts (Spec-X9-Decay-Semantics).\nReturns 202 with job_id for larger scopes; poll GET /v1/decay/jobs/:job_id.\nNote: dry_run is always synchronous per Spec-X9-Decay-Semantics.",
        "operationId": "decay_sweep_v1_decay_sweep_post",
        "parameters": [
          {
            "description": "Report what would be decayed without writing",
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": false,
              "description": "Report what would be decayed without writing",
              "title": "Dry Run",
              "type": "boolean"
            }
          },
          {
            "description": "Restrict sweep to one scope",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict sweep to one scope",
              "title": "Scope"
            }
          },
          {
            "description": "Expire non-expiring facts older than N seconds (0 = all)",
            "in": "query",
            "name": "ttl_seconds",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expire non-expiring facts older than N seconds (0 = all)",
              "title": "Ttl Seconds"
            }
          },
          {
            "description": "Expire active facts below this confidence",
            "in": "query",
            "name": "min_confidence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1.0,
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expire active facts below this confidence",
              "title": "Min Confidence"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Decay Sweep V1 Decay Sweep Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decay Sweep",
        "tags": [
          "decay"
        ]
      }
    },
...
  }
}